+++ tests/CustomException.php
@@ -10,23 +10,17 @@ class CustomException extends \Exception implements SolutionInterface
     public function getSolution(): Solution
     {
         return Solution::createSolution('My Solution')
-            ->setDescription('description')
+            ->setDescription('An custom description to exception')
             ->setDocs('https://google.com', 'See more');
     }
 
-    /**
-     * @param int $line
-     */
-    public function setLine(int $line)
+    /* public function setLine(int $line)
     {
         $this->line = $line;
     }
 
-    /**
-     * @param string $file
-     */
     public function setFile(string $file)
     {
         $this->file = $file;
-    }
+    } */
 }